requirements.txt
tensorflow-cpu
opencv-python
numpy
matplotlib
scikit-learn
imutils
